backend/pyproject.toml
[project]
name = "homebound-backend"
version = "0.1.0"
description = "Homebound API (FastAPI + SQLAlchemy)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "SQLAlchemy>=2.0",
  "aiosqlite>=0.19",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "itsdangerous>=2.2",
  "email-validator>=2.1",
  "PyJWT>=2.9",
  "httpx>=0.27",
  "certifi>=2024.0.0",
  "apscheduler>=3.10",
  "twilio>=9.0",
  "resend>=2.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "mypy>=1.11",
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "types-email-validator",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
